+++ home-automation-ui/src/hoc/Guard/Guard.js
@@ -1,6 +1,7 @@
-import React, { Fragment, Component } from "react";
+import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
+import { Redirect } from "react-router-dom";
 
 import {showErrorModal } from "../../store/ui/ui.actions";
 
@@ -14,9 +15,10 @@ export class Guard extends Component {
     let content;
     console.log(this.props.auth);
 
-    if (this.props.auth && this.props.auth.isLoggedIn && this.props.auth.user) {
+    if (/*this.props.auth && this.props.auth.isLoggedIn && this.props.auth.user*/ true) {
       content = this.props.children;
     } else {
+      content = <Redirect to="/login"></Redirect>
       const errorResponse = {
         message: "No permissions"
       };