+++ home-automation-ui/changes.md
@@ -10,3 +10,9 @@
 - Added ```Buzzer``` control.
 - Added ```Guard``` hoc component to restrict access.
 - Added ```auth``` store to handle authentification.
+
+10.03.2023
+- Added Input Component.
+- Added Login Page and Login Functionality.
+- Changed name in Header and Sidedrawer.
+- Added Logout to Header and Sidedrawer and Logout Functionality.